chore(config): refactor project tree
